+++ src/reason-parser/reason_pprint_ast.ml
@@ -162,7 +162,7 @@ let expression_immediate_extension_sugar x =
   | Some (name, expr) ->
     match expr.pexp_desc with
     | Pexp_for _ | Pexp_while _ | Pexp_ifthenelse _ | Pexp_function _
-    | Pexp_newtype _ | Pexp_try _ | Pexp_match _ (* | Pexp_letmodule _ *) ->
+    | Pexp_newtype _ | Pexp_try _ | Pexp_match _ ->
       (Some name, expr)
     | _ -> (None, x)
 
@@ -6236,8 +6236,9 @@ let printer = object(self:'self)
         | ([], Pexp_letop _) -> false
         | ([], Pexp_sequence _) -> false
         | ([], Pexp_letmodule _) -> false
-        | ([], Pexp_open (me, e)) ->
-          me.popen_override == Fresh && self#isSeriesOfOpensFollowedByNonSequencyExpression e
+        | ([], Pexp_open ({ popen_override; popen_expr = { pmod_desc = Pmod_ident _; _ }; _ }, e)) ->
+          popen_override == Fresh && self#isSeriesOfOpensFollowedByNonSequencyExpression e
+        | ([], Pexp_open _) -> false
         | ([], Pexp_letexception _) -> false
         | ([], Pexp_extension ({txt}, _)) -> txt = "mel.obj"
         | _ -> true
@@ -6478,7 +6479,7 @@ let printer = object(self:'self)
                     (atom (".")))
                   (self#formatNonSequencyExpression ~parent:x e))
             else
-              Some  (makeLetSequence (self#letList e))
+              Some (makeLetSequence (self#letList x))
         | Pexp_send (e, s) ->
           let needparens = match e.pexp_desc with
             | Pexp_apply (ee, _) ->
@@ -7812,10 +7813,9 @@ let printer = object(self:'self)
             self#attach_std_item_attrs binding.pmb_attributes module_binding
         | Pstr_open od ->
             self#attach_std_item_attrs od.popen_attributes @@
-            makeList ~postSpace:true [
-              atom ("open" ^ (override od.popen_override));
-              self#moduleExpressionToFormattedApplicationItems od.popen_expr;
-            ]
+            label ~space:true
+              (atom ("open" ^ (override od.popen_override)))
+              (self#moduleExpressionToFormattedApplicationItems od.popen_expr)
         | Pstr_modtype x ->
             let name = atom x.pmtd_name.txt in
             let letPattern = makeList ~postSpace:true [atom "module type"; name; atom "="] in