+++ test/modules.t/run.t
@@ -706,4 +706,24 @@ Format modules
   module%foo rec X: Y = {
     let x = 1;
   };
+  
+  let f = () => {
+    open {
+      let x = 1;
+    };
+  
+    ();
+  };
+  
+  let f = () => {
+    open {
+      let x = 1;
+    };
+  
+    ();
+  };
+  
+  open {
+    let x = 1;
+  };
 /* From http://stackoverflow.com/questions/1986374/  higher-order-type-constructors-and-functors-in-ocaml */